break out common functions into helpers
