+++ acceptance_tests/utilities/rh_endpoint_client.py
@@ -12,7 +12,7 @@ def is_exception_http_unauthorized(ex: BaseException) -> bool:
 
 
 # Retry on 401 errors as the UAC may not have been ingested into RH before the first attempt
-@retry(retry=retry_if_exception(is_exception_http_unauthorized), wait=wait_fixed(1), stop=stop_after_delay(10),
+@retry(retry=retry_if_exception(is_exception_http_unauthorized), wait=wait_fixed(1), stop=stop_after_delay(30),
        reraise=True)
 def post_to_launch_endpoint(uac: str) -> requests.Response:
     response = requests.post(f'{Config.RH_UI_URL}en/start/', data={'uac': uac}, allow_redirects=False)
